+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.4.1
 - Fix PHP warning: fatal error when end date < start date in AddDropBreakdownTime.php & CategoryBreakdownTime.php
 - Fix search discipline numeric field for "0" value in Widget.php
 - SQL case-insensitive username uniqueness check in Student.php, User.php & ImportUsers.fnc.php
+- Deprecate `$extra['second_col']` & `$extra['extra_search']`, use `$extra['search']` instead in ReferralLog.php, ClassSearchWidget.fnc.php, Search.inc.php & Export.php
 
 Changes in 12.4
 ---------------